SHOP/requirements.txt
asgiref==3.7.2
Django==5.0.3
djangorestframework==3.15.0
pillow==10.2.0
sqlparse==0.4.4
